pyproject.toml
[tool.poetry]
name = "drydock"
version = "0.1.0"
description = "Experiment in running ci pipelines using docker containers"
authors = ["zanderhavgaard <zander@havgaard.dk>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
docker = "^5.0.3"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
